# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student APP API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get a list of all students",
                "operationId": "4b649fd3816820e7d6ece134d72a7f52",
                "parameters": [
                    {
                        "name": "includeClassroom",
                        "in": "query",
                        "description": "Include the classroom relationship",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "includeDegree",
                        "in": "query",
                        "description": "Include the degree relationship",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StudentResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create a new student",
                "operationId": "20b07e0eabfb966bf0029eef43b01b6e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "classroom_id",
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "gender"
                                ],
                                "properties": {
                                    "classroom_id": {
                                        "description": "The ID of the classroom the student belongs to",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "first_name": {
                                        "description": "The first name of the student",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "description": "The last name of the student",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "description": "The email address of the student",
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "gender": {
                                        "description": "The gender of the student",
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ],
                                        "example": "male"
                                    },
                                    "date_of_birth": {
                                        "description": "The date of birth of the student (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2004-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Student created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/StudentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        },
                                        "errors": {
                                            "$ref": "#/components/schemas/StoreStudentRequest"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreStudentRequest": {},
...
            "StudentResource": {}
        }
    }
}